backport make, ahrs separation, bugfixes

+++ src/main/io/flashfs.c
@@ -70,8 +70,8 @@ static DMA_DATA_ZERO_INIT uint8_t flashWriteBuffer[FLASHFS_WRITE_BUFFER_SIZE];
  * The tail is advanced once a write is complete up to the location behind head. The tail is advanced
  * by a callback from the FLASH write routine. This prevents data being overwritten whilst a write is in progress.
  */
-static uint8_t bufferHead = 0;
-static volatile uint8_t bufferTail = 0;
+static uint16_t bufferHead = 0;
+static volatile uint16_t bufferTail = 0;
 
 /* Track if there is new data to write. Until the contents of the buffer have been completely
  * written flashfsFlushAsync() will be repeatedly called. The tail pointer is only updated
